Move relevant tests to shared library
