+++ common/go-utils/src/inv.rs
@@ -85,7 +85,6 @@ impl UpstreamInventory<GoVersion> for Inventory<GoVersion> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::hash::{BuildHasher, RandomState};
 
     fn create_artifact() -> Artifact<GoVersion> {
         Artifact::<GoVersion> {
@@ -106,28 +105,5 @@ mod tests {
         let artifact = create_artifact();
 
         assert_eq!("Go 1.7.2 (linux-x86_64)", artifact.to_string());
-    }
-
-    #[test]
-    fn test_artifact_hash_implementation() {
-        let artifact = create_artifact();
-
-        let state = RandomState::new();
-        assert_eq!(
-            state.hash_one(&artifact.checksum.value),
-            state.hash_one(&artifact)
-        );
-    }
-
-    #[test]
-    fn test_artifact_serialization() {
-        let artifact = create_artifact();
-        let serialized = toml::to_string(&artifact).unwrap();
-        assert!(serialized
-            .contains("sha256:abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"));
-        assert_eq!(
-            artifact,
-            toml::from_str::<Artifact<GoVersion>>(&serialized).unwrap()
-        );
     }
 }
